requirements.txt
Pillow==2.7.0
argparse==1.2.1
wsgiref==0.1.2
